Add files via upload
This update adds a new menu option 5 – View History that shows deposits, withdrawals, and balance checks with timestamps for the current session. It improves usability by letting users review recent actions, similar to a real ATM. The code was also cleaned by shortening comments to keep it simpler and easier to read.